refactor(chat): isolate SSE transport
